Refactor views
