+++ enhydris/views/__init__.py
@@ -1,0 +1,12 @@
+from .station import StationDetail, StationEdit, StationList
+from .timeseries import DownloadData
+from .timeseries_group import OldTimeseriesDetailRedirectView, TimeseriesGroupDetail
+
+__all__ = (
+    "StationDetail",
+    "StationEdit",
+    "StationList",
+    "DownloadData",
+    "OldTimeseriesDetailRedirectView",
+    "TimeseriesGroupDetail",
+)
